reader/requirements.txt
httpx==0.27.2
mfrc522==0.0.7   
RPi.GPIO==0.7.1  
